requirements.txt
github3.py
requests
beautifulsoup4
sendgrid
python-dateutil
logging
sqlalchemy
pymysql
datetime
pyyaml